LPS-66799 Update the base classes - cleaner, more separated responsibilities, usage of generics
