+++ modules/apps/web-experience/export-import/export-import-service/build.gradle
@@ -3,7 +3,6 @@ targetCompatibility = "1.8"
 
 dependencies {
 	provided group: "biz.aQute.bnd", name: "biz.aQute.bndlib", version: "3.1.0"
-	provided group: "com.liferay", name: "com.liferay.exportimport.api", version: "2.1.0"
 	provided group: "com.liferay", name: "com.liferay.petra.xml", version: "1.0.0"
 	provided group: "com.liferay", name: "com.liferay.portal.background.task.api", version: "2.0.0"
 	provided group: "com.liferay", name: "com.liferay.portal.background.task.service", version: "2.0.0"
@@ -20,4 +19,5 @@ dependencies {
 	provided group: "org.osgi", name: "org.osgi.core", version: "6.0.0"
 	provided group: "org.osgi", name: "org.osgi.service.component", version: "1.3.0"
 	provided group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.3.0"
+	provided project(":apps:web-experience:export-import:export-import-api")
 }
