+++ modules/apps/web-experience/export-import/export-import-service/src/main/java/com/liferay/exportimport/lar/BaseStagedModelDataHandler.java
@@ -15,7 +15,6 @@
 package com.liferay.exportimport.lar;
 
 import com.liferay.exportimport.content.processor.ExportImportContentProcessor;
-import com.liferay.exportimport.content.processor.ExportImportContentProcessorRegistryUtil;
 import com.liferay.exportimport.kernel.lar.ExportImportDateUtil;
 import com.liferay.exportimport.kernel.lar.ExportImportProcessCallbackRegistryUtil;
 import com.liferay.exportimport.kernel.lar.ExportImportThreadLocal;
@@ -147,14 +146,11 @@ public abstract class BaseStagedModelDataHandler<T extends StagedModel>
 			portletDataContext, stagedModel);
 	}
 
+	@Deprecated
 	protected ExportImportContentProcessor getExportImportContentProcessor(
 		Class<T> clazz) {
 
-		ExportImportContentProcessor exportImportContentProcessor =
-			ExportImportContentProcessorRegistryUtil.
-				getExportImportContentProcessor(clazz.getName());
-
-		return exportImportContentProcessor;
+		return null;
 	}
 
 	protected StagedModelRepository<T> getStagedModelRepository() {
